requirements.txt
aniso8601==9.0.1
antiorm==1.2.1
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
Charm-Crypto==0.50
charset-normalizer==3.4.0
click==8.1.7
ConfigArgParse==1.7
db==0.1.1
db-sqlite3==0.0.1
debugpy==1.8.9
dnspython==2.7.0
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
flask-restx==1.3.0
flask-swagger-ui==4.11.1
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
hiredis==3.1.0
hvac==2.3.0
hypothesis==6.119.4
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
locust==2.32.6
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
pluggy==1.5.0
psutil==6.1.1
psycopg2-binary==2.9.10
pycryptodome==3.21.0
pyparsing==2.4.0
pytest==8.3.3
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
rstr==3.2.2
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.37
tomli==2.1.0
typing_extensions==4.12.2
urllib3==2.2.3
watchdog==6.0.0
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
